Bump sqlparse from 0.4.1 to 0.4.4
The new version of sqlparse fixes two vulnerabilities, but changes how the IN
operator is treated within the WHERE clause. This commit ensures the
rewriting for key-value stores, part of the tests, works despite the change.

Other few minor fixes include: clarification of the rewriting limitations,
removal of unsed methods, and properly rewriting query for key-value stores
when no WHERE clause is present.